# specifications/architecture/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DigitalBurj Logistics OS API",
    "version": "1.0.0",
    "description": "Versioned REST contract. Money = decimal strings. Commands are idempotent via Idempotency-Key; mutable aggregates use If-Match."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "SERVICE_UNAVAILABLE",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "TENANT_REQUIRED",
              "NOT_FOUND",
              "VALIDATION_FAILED",
              "VERSION_CONFLICT",
              "IDEMPOTENCY_KEY_REUSED",
              "IDEMPOTENCY_KEY_REQUIRED",
              "INVALID_STATE_TRANSITION",
              "DUPLICATE",
              "ACCOUNTING_PERIOD_CLOSED",
              "PERIOD_NOT_FOUND",
              "INSUFFICIENT_STOCK",
              "STOCK_ON_HOLD",
              "RELEASE_EVIDENCE_REQUIRED",
              "SEPARATION_OF_DUTIES",
              "OVER_ALLOCATION",
              "JOB_NOT_CLOSABLE",
              "CURRENCY_MISMATCH",
              "DOCUMENT_NOT_CLEAN",
              "WEBHOOK_SIGNATURE_INVALID",
              "AI_TOOL_DENIED",
              "AI_BUDGET_EXCEEDED",
              "BOOKING_OUTCOME_UNKNOWN",
              "QUALIFICATION_REQUIRED",
              "RATE_LIMITED",
              "INTERNAL"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/v1/work/calendars": {
      "get": {
        "operationId": "listWorkCalendars",
        "tags": [
          "Work"
        ],
        "summary": "List immutable calendar versions",
        "description": "Permission: work.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWorkCalendar",
        "tags": [
          "Work"
        ],
        "summary": "Publish a business calendar version",
        "description": "Permission: work.configure",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_-]{1,50}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "calendar": {
                    "type": "object",
                    "properties": {
                      "timezone": {
                        "type": "string",
                        "maxLength": 80
                      },
                      "weekdays": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 7
                        },
                        "minItems": 1,
                        "maxItems": 7
                      },
                      "startMinute": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 1439
                      },
                      "endMinute": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 1440
                      },
                      "holidays": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        "maxItems": 366,
                        "default": []
                      }
                    },
                    "required": [
                      "timezone",
                      "weekdays",
                      "startMinute",
                      "endMinute"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "key",
                  "name",
                  "calendar"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/templates": {
      "get": {
        "operationId": "listWorkTemplates",
        "tags": [
          "Work"
        ],
        "summary": "List scoped workflow templates",
        "description": "Permission: work.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWorkTemplate",
        "tags": [
          "Work"
        ],
        "summary": "Publish an immutable dependency template",
        "description": "Permission: work.configure",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalEntityId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "branchId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid",
                    "default": null
                  },
                  "key": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_-]{1,50}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "calendarId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "steps": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_-]{1,50}$"
                        },
                        "title": {
                          "type": "string",
                          "minLength": 2,
                          "maxLength": 120
                        },
                        "department": {
                          "type": "string",
                          "minLength": 2,
                          "maxLength": 120
                        },
                        "slaMinutes": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 43200
                        },
                        "complexity": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 100,
                          "default": 1
                        },
                        "dependsOn": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "maxLength": 51
                          },
                          "maxItems": 20,
                          "default": []
                        }
                      },
                      "required": [
                        "key",
                        "title",
                        "department",
                        "slaMinutes"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 30
                  }
                },
                "required": [
                  "legalEntityId",
                  "key",
                  "name",
                  "calendarId",
                  "steps"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/instances": {
      "post": {
        "operationId": "startWorkTemplate",
        "tags": [
          "Work"
        ],
        "summary": "Start template work with dependency gates",
        "description": "Permission: work.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "templateId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "assigneeMembershipId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid",
                    "default": null
                  }
                },
                "required": [
                  "templateId",
                  "title"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/items": {
      "get": {
        "operationId": "listWorkItems",
        "tags": [
          "Work"
        ],
        "summary": "Department queue with entity and branch scopes",
        "description": "Permission: work.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 120
            },
            "required": false,
            "name": "department",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "blocked",
                "open",
                "in_progress",
                "waiting",
                "done",
                "cancelled"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "mine",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWorkItem",
        "tags": [
          "Work"
        ],
        "summary": "Create assigned work with a business calendar SLA",
        "description": "Permission: work.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalEntityId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "branchId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid",
                    "default": null
                  },
                  "title": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 4000
                  },
                  "department": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "calendarId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "slaMinutes": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 43200
                  },
                  "complexity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100,
                    "default": 1
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "normal",
                      "high",
                      "critical"
                    ],
                    "default": "normal"
                  },
                  "assigneeMembershipId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid",
                    "default": null
                  }
                },
                "required": [
                  "legalEntityId",
                  "title",
                  "description",
                  "department",
                  "calendarId",
                  "slaMinutes"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/items/{id}": {
      "get": {
        "operationId": "getWorkItem",
        "tags": [
          "Work"
        ],
        "summary": "Work detail and audit history",
        "description": "Permission: work.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/items/{id}/transition": {
      "post": {
        "operationId": "transitionWorkItem",
        "tags": [
          "Work"
        ],
        "summary": "Start pause resume complete or cancel work",
        "description": "Permission: work.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "start",
                      "pause",
                      "resume",
                      "complete",
                      "cancel"
                    ]
                  },
                  "note": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "action",
                  "note"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/items/{id}/assign": {
      "post": {
        "operationId": "assignWorkItem",
        "tags": [
          "Work"
        ],
        "summary": "Reassign to qualified scoped staff",
        "description": "Permission: work.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assigneeMembershipId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "assigneeMembershipId",
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/directory": {
      "get": {
        "operationId": "getWorkDirectory",
        "tags": [
          "Work"
        ],
        "summary": "Scoped entities branches and active staff directory",
        "description": "Permission: work.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/capacity": {
      "post": {
        "operationId": "setWorkCapacity",
        "tags": [
          "Work"
        ],
        "summary": "Configure department capacity and availability",
        "description": "Permission: work.configure",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalEntityId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "branchId": {
                    "type": "string",
                    "nullable": true,
                    "format": "uuid",
                    "default": null
                  },
                  "membershipId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "department": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "capacityPoints": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000
                  },
                  "available": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "legalEntityId",
                  "membershipId",
                  "department",
                  "capacityPoints"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/workload": {
      "get": {
        "operationId": "getWorkload",
        "tags": [
          "Work"
        ],
        "summary": "Scoped capacity utilization and overdue work",
        "description": "Permission: work.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work/check-slas": {
      "post": {
        "operationId": "checkWorkSlas",
        "tags": [
          "Work"
        ],
        "summary": "Record newly breached SLAs exactly once",
        "description": "Permission: work.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/work/handovers/{id}/acknowledge": {
      "post": {
        "operationId": "acknowledgeHandover",
        "tags": [
          "Work"
        ],
        "summary": "Recipient acknowledges and atomically receives work",
        "description": "Permission: work.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "note"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}